api/requirements.txt
fastapi[standard]
PyJWT